+++ gui/src/cel_shaded_generator_gui/tabs/__init__.py
@@ -1,5 +1,11 @@
 from .animation_tab import MangaAnimationTab as MangaAnimationTab
 from .colorization_tab import MangaColorizationTab as MangaColorizationTab
 from .puppeteering_tab import MangaPuppeteeringTab as MangaPuppeteeringTab
+from .reference_coloring_tab import ReferenceColoringTab as ReferenceColoringTab
 
-__all__ = ["MangaColorizationTab", "MangaAnimationTab", "MangaPuppeteeringTab"]
+__all__ = [
+    "MangaColorizationTab",
+    "MangaAnimationTab",
+    "MangaPuppeteeringTab",
+    "ReferenceColoringTab",
+]